moved sampler to lib folder, fixed cofig
