+++ lib/scripts/sampler.js
